portfolio_manager.py
import yfinance as yf
from decimal import Decimal

class PortfolioManager:
    def __init__(self):
        pass

    def rebalance(self, portfolio_weights, accounts, current_allocations):
        # Placeholder implementation
        return []

    def get_price(self, ticker):
        # Fetch price from yfinance
        return Decimal(str(yf.Ticker(ticker).info['regularMarketPrice']))

    def calculate_total_value(self, current_allocations, accounts):
        # Calculate total portfolio value including idle cash
        total_value = sum(Decimal(account['Idle_Cash']) for account in accounts)
        for allocation in current_allocations:
            ticker = allocation['Ticker']
            shares = Decimal(allocation['Shares'])
            price = self.get_price(ticker)
            total_value += shares * price
        return total_value

    def calculate_target_allocations(self, portfolio_weights, total_value):
        # Calculate target allocations based on portfolio weights and total value
        return [
            {
                'Ticker': weight['Ticker'],
                'TargetValue': Decimal(weight['Cash_Weight']) * total_value
            }
            for weight in portfolio_weights
        ]

    def generate_orders(self, current_allocations, target_allocations, accounts):
        # Generate buy/sell orders to reach target allocations
        # This is a placeholder and needs to be implemented
        return []
